% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cultivars.R
\docType{data}
\name{cultivars}
\alias{cultivars}
\title{Cultivars Data}
\format{
a vector with 50 records
}
\source{
Karpenstein-Machen, B. Honermeier, and F. Hartmann. Produktion Aktuell, Triticale.
DLG Verlag, Frankfurt, Germany, 1994.
}
\usage{
cultivars
}
\description{
The cultivars dataset is taken from Karpenstein-Machen et al. (1994) and
Karpenstein-Machan and Maschka (1996). It has also been analyzed by Piepho (2000).
}
\keyword{datasets}
