% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse.R
\docType{data}
\name{pulse}
\alias{pulse}
\title{Pulse Rate Data}
\format{
a vector with 50 records
}
\source{
Hand, F. Daly, A. Lunn, K. McConway, and E. Ostrowsky. A Handbook of
Small Data Sets. Chapman and Hall, London, UK, 1994.
}
\usage{
pulse
}
\description{
At a hospital the pulse rates of 50 patients were measured in beats per minute.
The data are taken from Hand et al. (1994)
}
\keyword{datasets}
