% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stGOF.R
\name{stGOF}
\alias{stGOF}
\title{Smooth Tests}
\usage{
stGOF(
  formula,
  data = NA,
  order = NULL,
  method = "MLE",
  rescale = FALSE,
  B = NULL,
  max.order = 0,
  horizon = "",
  criterion = "",
  output = TRUE
)
}
\arguments{
\item{formula}{A formula is used to provide the data and the distribution.
And it should be written as \code{Data ~ distr}. The distribution to be tested,
and must be one of "unif" (default), "pois", "logar", "exp", "norm", "negb",
"ZIP", "laplace", "extrval", "geom", "ZTP",
"logis", "genpareto", "betab" or "gamma"}

\item{data}{A numeric vector of sample observations}

\item{order}{The order of the test}

\item{method}{Indicates the method for parameter estimation and must be one
of "MLE" (default), or "MME"}

\item{rescale}{If \code{rescale}=TRUE the empirical variance is used to
rescale the components; if \code{rescale}=FALSE (default) the theoretical
variance under the null hypothesis is used.}

\item{B}{The number of bootstrap runs for p-value calculation. When
\code{B}=NULL the asymptotic chi-squared distribution is used for p-value
calculation.}

\item{max.order}{The maximal order of the test that can be selected by the
model selection criterion}

\item{horizon}{\code{horizon}="order" (default) refers to order-selection,
and \code{horizon}="subset" refers to subset selection}

\item{criterion}{A character specifying the model selection criterion:
"AIC" (default), or "BIC"}

\item{output}{Logical item; if TRUE (default) an extensive output of the smooth test is given, otherwise, no output is given.}
}
\value{
A list with components (in case of fixed order)  :

-\code{statistics}{ a numeric vector containing the (rescaled) components. The last position in the vector contains
 the smooth test statistic.}

-\code{p.value}{ a numeric vector containing the p-values corresponding to \code{statistics}}

-\code{par.est}{ a numeric vector containing the nuisance parameter estimates}

-\code{comp}{ a numeric vector containing the (unscaled) components (this is exactly equal to \code{statistics} without the last entry).}

-\code{Sigma}{ the theoretical variance covariance matrix of the (unscaled) components, calculated under the null hypothesis}

-\code{EVar}{ a numeric vector containing the empirical variances of the (unscaled) components}

A list with components (in case of data-driven test):

-\code{stat}{ the test statistic}

-\code{model}{ a vector with the selected orders}

-\code{p.value}{ the p-value}

-\code{par.est}{ a numeric vector containing the nuisance parameter estimates}
}
\description{
Performs the smooth test for the one-sample goodness-of-fit
problem as described by Rayner et al.(2009). Both simple and composite null
hypotheses can be tested. The maximum likelihood (MLE) and the method of
moments (MME) methods for nuisance parameter estimation are implemented.
Function returns the score statistics of Smooth test result.
}
\details{
{This function performs a smooth test of goodness of fit for both
simple and composite null hypotheses. The nuisance parameters can be estimated
by means of maximum likelihood (\code{method}="MLE) or method of moments}
(\code{method}="MME). The function also performs the individual component tests.
The p-values are computed using the asymptotic chi-squared approximation
(\code{B}=NULL) or using a parametric bootstrap approximation (\code{B}>0).
The \code{stGOF} function works with a fixed order and in a data-driven smooth
test. In the data-driven case, the components are selected from a horizon that is specified by
a maximal order (\code{max.order}), a model selection criterion
(\code{criterion}) and a type of selection (\code{horizon}="order" for order
selection, and \code{horizon}="subset" for subset selection).
}
\examples{
\code{stGOF(PRG ~ unif, PRG, order = 4, method = "MLE")}
\code{stGOF(PCB ~ norm, PCB, order = 4, method = "MLE")}
\code{stGOF(PCB ~ norm, PCB, order = 5, method = "MLE", B = 1000, rescale = T)}
\code{stGOF(PCB ~ norm, PCB, method = "MLE", B = 1000, max.order = 7, horizon="order", criterion="AIC")}
\code{stGOF(PCB ~ norm, PCB, method = "MME", B = 1000, max.order = 7, horizon="subset", criterion="BIC")}
\code{stGOF(pulse ~ pois, PCB, method = "MME", B = 1000, max.order = 8, horizon="subset", criterion="BIC")}

}
\references{
Rayner, J.C.W., Thas, O. and Best, D.J. (2009). Smooth Tests of Goodness-of-fit. Wiley, Singapore.
}
\author{
Mahmoud ABU AZOUM, and Prof. Olivier Thas
}
\keyword{smooth}
\keyword{test}
